ネズミAI完成。コリジョンが横スライドする攻撃モーションに未対応。

+++ mmo/Assets/Terasurware/Classes/Editor/EnemyData_importer.cs
@@ -52,8 +52,8 @@ public class EnemyData_importer : AssetPostprocessor {
 					cell = row.GetCell(5); p.MagicDefense = (int)(cell == null ? 0 : cell.NumericCellValue);
 					cell = row.GetCell(6); p.MoveSpeed = (float)(cell == null ? 0 : cell.NumericCellValue);
 					cell = row.GetCell(7); p.TrackingSpeed = (float)(cell == null ? 0 : cell.NumericCellValue);
-					cell = row.GetCell(8); p.ActionInterval = (int)(cell == null ? 0 : cell.NumericCellValue);
-					cell = row.GetCell(9); p.FieldOfView = (int)(cell == null ? 0 : cell.NumericCellValue);
+					cell = row.GetCell(8); p.ActionInterval = (float)(cell == null ? 0 : cell.NumericCellValue);
+					cell = row.GetCell(9); p.FieldOfView = (float)(cell == null ? 0 : cell.NumericCellValue);
 					cell = row.GetCell(10); p.ViewDistance = (float)(cell == null ? 0 : cell.NumericCellValue);
 					cell = row.GetCell(11); p.ActionDistance = (float)(cell == null ? 0 : cell.NumericCellValue);
 						s.list.Add (p);
